diabetes-cgm-project/requirements.txt
pandas
numpy
matplotlib
seaborn
singer-python